Move ProcessExecutor into its own module.
