+++ labtech/runners/process.py
@@ -1,246 +1,41 @@
 from __future__ import annotations
 
-import functools
 import logging
 import multiprocessing
-import os
 import signal
 import sys
 from abc import ABC, abstractmethod
-from dataclasses import dataclass, field
-from enum import StrEnum, auto
-from itertools import count
+from dataclasses import dataclass
 from logging.handlers import QueueHandler
 from queue import Empty
-from threading import Thread
 from typing import TYPE_CHECKING, cast
 from uuid import uuid4
 
 import psutil
 
-from labtech.exceptions import RunnerError, TaskDiedError
+from labtech.exceptions import RunnerError
 from labtech.monitor import get_process_info
 from labtech.tasks import get_direct_dependencies
 from labtech.types import Runner, RunnerBackend
 from labtech.utils import LoggerFileProxy, get_supported_start_methods, is_interactive, logger
 
+from ._process_executor import ProcessExecutor
 from .base import run_or_load_task
 
 if TYPE_CHECKING:
     from collections.abc import Callable, Iterator, Sequence
     from multiprocessing.context import BaseContext, SpawnContext
     from queue import Queue
-    from typing import Any
     from uuid import UUID
 
     from labtech.types import LabContext, ResultMeta, ResultsMap, Storage, Task, TaskMonitorInfo, TaskResult
+
+    from ._process_executor import ExecutorFuture
 
     if sys.platform != 'win32':
         from multiprocessing.context import ForkContext
     else:
         ForkContext = BaseContext
-
-
-class FutureStateError(Exception):
-    pass
-
-
-class FutureState(StrEnum):
-    PENDING = auto()
-    CANCELLED = auto()
-    FINISHED = auto()
-
-
-@dataclass
-class Future:
-    """Representation of a result to be returned in the future by a runner.
-
-    A Future's state transitions between states according to the following
-    finite state machine:
-
-    * A Future starts in a PENDING state
-    * A PENDING Future can be transitioned to FINISHED by calling
-      set_result() or set_exception()
-    * Any Future can be transitioned to CANCELLED by calling cancel()
-    * result() can only be called on a FINISHED Future, and it will either
-      return the result set by set_result() or raise the exception set by
-      set_exception()
-
-    """
-    # Auto-incrementing ID (does not need to be process-safe because
-    # all futures are generated in the main process):
-    id: int = field(default_factory=count().__next__, init=False)
-    _state: FutureState = FutureState.PENDING
-    _ex: BaseException | None = None
-    _result: Any | None = None
-
-    def __eq__(self, other: object) -> bool:
-        if not isinstance(other, self.__class__):
-            return False
-        return id(self.id) == id(other.id)
-
-    def __hash__(self) -> int:
-        return hash(self.id)
-
-    @property
-    def done(self) -> bool:
-        return self._state in {FutureState.FINISHED, FutureState.CANCELLED}
-
-    @property
-    def cancelled(self) -> bool:
-        return self._state == FutureState.CANCELLED
-
-    def set_result(self, result: Any):
-        if self.done:
-            raise FutureStateError(f'Attempted to set a result on a {self._state} future.')
-        self._result = result
-        self._state = FutureState.FINISHED
-
-    def set_exception(self, ex: BaseException):
-        if self.done:
-            raise FutureStateError(f'Attempted to set an exception on a {self._state} future.')
-        self._ex = ex
-        self._state = FutureState.FINISHED
-
-    def cancel(self):
-        self._state = FutureState.CANCELLED
-
-    def result(self) -> Any:
-        if self._state != FutureState.FINISHED:
-            raise FutureStateError(f'Attempted to get result from a {self._state} future.')
-        if self._ex is not None:
-            raise self._ex
-        return self._result
-
-
-def split_done_futures(futures: Sequence[Future]) -> tuple[list[Future], list[Future]]:
-    done_futures = []
-    not_done_futures = []
-    for future in futures:
-        if future.done:
-            done_futures.append(future)
-        else:
-            not_done_futures.append(future)
-    return (done_futures, not_done_futures)
-
-
-def _subprocess_target(*, future_id: int, thunk: Callable[[], Any], result_queue: Queue) -> None:
-    try:
-        result = thunk()
-    except BaseException as ex:
-        result_queue.put((future_id, ex))
-    else:
-        result_queue.put((future_id, result))
-
-
-class ProcessExecutor:
-
-    def __init__(self, mp_context: BaseContext, max_workers: int | None):
-        self.mp_context = mp_context
-        self.max_workers = (os.cpu_count() or 1) if max_workers is None else max_workers
-        self._pending_future_to_thunk: dict[Future, Callable[[], Any]] = {}
-        self._running_id_to_future_and_process: dict[int, tuple[Future, multiprocessing.Process]] = {}
-        # Use a Manager().Queue() to be able to share with subprocesses
-        self._result_queue: Queue = multiprocessing.Manager().Queue(-1)
-
-    def _start_processes(self):
-        """Start processes for the oldest pending futures to bring
-        running process count up to max_workers."""
-        start_count = max(0, self.max_workers - len(self._running_id_to_future_and_process))
-        futures_to_start = list(self._pending_future_to_thunk.keys())[:start_count]
-        for future in futures_to_start:
-            thunk = self._pending_future_to_thunk[future]
-            del self._pending_future_to_thunk[future]
-            process = self.mp_context.Process(
-                target=_subprocess_target,
-                kwargs=dict(
-                    future_id=future.id,
-                    thunk=thunk,
-                    result_queue=self._result_queue,
-                ),
-            )
-            self._running_id_to_future_and_process[future.id] = (future, process)
-            process.start()
-
-    def submit(self, fn: Callable, /, *args, **kwargs) -> Future:
-        """Schedule the given fn to be called with the given *args and
-        **kwargs, and return a Future that will be updated with the
-        outcome of function call."""
-        future = Future()
-        self._pending_future_to_thunk[future] = functools.partial(fn, *args, **kwargs)
-        self._start_processes()
-        return future
-
-    def cancel(self) -> None:
-        """Cancel all pending futures."""
-        pending_futures = list(self._pending_future_to_thunk.keys())
-        for future in pending_futures:
-            future.cancel()
-            del self._pending_future_to_thunk[future]
-
-    def stop(self) -> None:
-        """Cancel all running futures and immediately terminate their execution."""
-        future_process_pairs = list(self._running_id_to_future_and_process.values())
-        for future, process in future_process_pairs:
-            process.terminate()
-            future.cancel()
-            del self._running_id_to_future_and_process[future.id]
-
-    def _consume_result_queue(self, *, timeout_seconds: float | None):
-        # Avoid race condition of a process finishing after we have
-        # consumed the result_queue by fetching process statuses
-        # before checking for process completion.
-        dead_process_futures = [
-            future for future, process in self._running_id_to_future_and_process.values()
-            if not process.is_alive()
-        ]
-
-        def _consume():
-            inner_timeout_seconds = timeout_seconds
-            while True:
-                try:
-                    future_id, result_or_ex = self._result_queue.get(True, timeout=inner_timeout_seconds)
-                except Empty:
-                    break
-
-                # Don't wait for the timeout on subsequent calls to
-                # self._result_queue.get()
-                inner_timeout_seconds = 0
-
-                future, _ = self._running_id_to_future_and_process[future_id]
-                del self._running_id_to_future_and_process[future_id]
-                if not future.done:
-                    if isinstance(result_or_ex, BaseException):
-                        future.set_exception(result_or_ex)
-                    else:
-                        future.set_result(result_or_ex)
-
-        # Consume the result queue in a thread so that it is not
-        # interrupt by KeyboardInterrupt, which can result in us not
-        # fully processing a completed result. Despite the fact we are
-        # using subprocesses, starting a thread at this point should
-        # be safe because we will not start any subprocesses while
-        # this is running?
-        consumer_thread = Thread(target=_consume)
-        consumer_thread.start()
-        consumer_thread.join()
-
-        # If any processes have died without the future being
-        # cancelled or finished, then set an exception for it.
-        for future in dead_process_futures:
-            if future.done:
-                continue
-            future.set_exception(TaskDiedError())
-            del self._running_id_to_future_and_process[future.id]
-
-    def wait(self, futures: Sequence[Future], *, timeout_seconds: float | None) -> tuple[list[Future], list[Future]]:
-        """Wait up to timeout_seconds or until at least one of the
-        given futures is done, then return a list of futures in a done
-        state and a list of futures in all other states."""
-        self._consume_result_queue(timeout_seconds=timeout_seconds)
-        # Having consumed completed results, start new processes
-        self._start_processes()
-        return split_done_futures(futures)
 
 
 class ProcessEvent:
@@ -325,7 +120,7 @@ class ProcessRunner(Runner, ABC):
         )
 
         self.results_map: dict[Task, TaskResult] = {}
-        self.future_to_task: dict[Future, Task] = {}
+        self.future_to_task: dict[ExecutorFuture, Task] = {}
 
     def _consume_log_queue(self):
         # See: https://docs.python.org/3/howto/logging-cookbook.html#logging-to-a-single-file-from-multiple-processes
@@ -446,9 +241,9 @@ class ProcessRunner(Runner, ABC):
 
     @abstractmethod
     def _submit_task(self, executor: ProcessExecutor, task: Task, task_name: str,
-                     use_cache: bool, process_event_queue: Queue, log_queue: Queue) -> Future:
+                     use_cache: bool, process_event_queue: Queue, log_queue: Queue) -> ExecutorFuture:
         """Should submit the execution of self._subprocess_func() on the given
-        task to the given executor and return the resulting Future.
+        task to the given executor and return the resulting ExecutorFuture.
 
         Sub-classes can use the implementation of this method to load
         or otherwise prepare context or dependency results for the task.
@@ -467,7 +262,7 @@ class SpawnProcessRunner(ProcessRunner):
         return multiprocessing.get_context('spawn')
 
     def _submit_task(self, executor: ProcessExecutor, task: Task, task_name: str,
-                     use_cache: bool, process_event_queue: Queue, log_queue: Queue) -> Future:
+                     use_cache: bool, process_event_queue: Queue, log_queue: Queue) -> ExecutorFuture:
         if is_interactive() and task.__class__.__module__ == '__main__':
             raise RunnerError(
                 (f'Unable to submit {task.__class__.__qualname__} tasks to '
@@ -567,7 +362,7 @@ class ForkProcessRunner(ProcessRunner):
         )
 
     def _submit_task(self, executor: ProcessExecutor, task: Task, task_name: str,
-                     use_cache: bool, process_event_queue: Queue, log_queue: Queue) -> Future:
+                     use_cache: bool, process_event_queue: Queue, log_queue: Queue) -> ExecutorFuture:
         return executor.submit(
             self._fork_subprocess_func,
             _subprocess_func=self._subprocess_func,
